requirement.txt
streamlit
streamlit-mic-recorder
openai-whisper
moviepy
pydub
scipy
audioop-lts